+++ Expressions/LiteralExpression.cs
@@ -61,7 +61,7 @@ namespace Exodrifter.Rumor.Expressions
 			}
 		}
 
-		public override Value Evaluate(Engine.Rumor rumor)
+		public override Value Evaluate(Scope scope, Bindings bindings)
 		{
 			return value;
 		}